requirements.txt
beautifulsoup4>=4.10.0
Pillow>=9.4.0
requests>=2.26.0
